app/requirements.txt
fastapi
pymongo
pydantic
uvicorn[standard]
python-dotenv
pydantic[email]